python-api/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
opencv-python-headless
numpy